LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/document.open-02-expected.txt


PASS document.open should redirect to window.open when called with three arguments
PASS document.open should throw when it has no window and is called with three arguments
